Children are not equal if location has no key
If a route is rendered with no key (this was seen when revisiting a site in a new window with a valid token) then the children were never equal and therefore it was rendering in an infinite loop resulting in the transition repeatedly fading out and back in. This fixes the problem.